♻️ Refactor : 기존에 구현한 brand toggle api 파일 리팩토링
